refactor: consolidate ApiError and ApiErrorCode into client.ts, delete apiError.ts
